No zero-points filter
Stopped filtering out zero-point assignments. It appears that some faculty are intentionally creating feedback-only zero-point assignments that need to be accounted for in their analytics. So, we'll just tally zero-point assignments and then maybe do a check to see if they are, for example complete/incomplete assignments (indicating that they are intentional). Assessing intentionality through data is hard.

Also improved the color-coding on the departmental summary for greater clarity.