Reflexive Canvas LTI
Bah. No point in supporting the old, crufty LTI structure. Kicking it over to the new, crufty LTI structure while trying to bring this back online. As of right now, the graphs aren't loading (PHPGraphLib seems not to properly autoload, and this is probably the impetus to ditch it) and the cron script has been moved so the crontab installation documentation/script needs to be updated.

There's a weirdness around the LTI Configuration XML generator that I haven't tracked down yet — `$toolbox->getGenerator()->saveXML()` seems to reliably think that it is calling `saveXML()` on `null` — which makes no sense, since `$toolbox->generator` is actually an instance of `Generator`. I've grafted in a fix on my development build right now that is NOT fit for primetime (line 802 in `vendor/smtech/reflexive-canvas-lti/src/Toolbox.php`):

```PHP
$this->getGenerator(); // initializes $this->generator
$this->generator->saveXML(); // so weird!
```